+++ client/group.js
@@ -1,57 +1,40 @@
-Group = function(router, options, parent) {
-  options = options || {};
+Group = class extends SharedGroup {
+  constructor(router, options, parent) {
+    super(router, options, parent);
 
-  if (options.prefix && !/^\/.*/.test(options.prefix)) {
-    var message = "group's prefix must start with '/'";
-    throw new Error(message);
+    options = options || {};
+
+    this.name = options.name;
+
+    this._triggersEnter = options.triggersEnter || [];
+    this._triggersExit = options.triggersExit || [];
+    this._subscriptions = options.subscriptions || Function.prototype;
+
+    if (this.parent) {
+      this.prefix = parent.prefix + this.prefix;
+
+      this._triggersEnter = parent._triggersEnter.concat(this._triggersEnter);
+      this._triggersExit = this._triggersExit.concat(parent._triggersExit);
+    }
   }
 
-  this._router = router;
-  this.prefix = options.prefix || '';
-  this.name = options.name;
-  this.options = options;
+  route(pathDef, options, group) {
+    options = options || {};
 
-  this._triggersEnter = options.triggersEnter || [];
-  this._triggersExit = options.triggersExit || [];
-  this._subscriptions = options.subscriptions || Function.prototype;
+    const triggersEnter = options.triggersEnter || [];
+    options.triggersEnter = this._triggersEnter.concat(triggersEnter);
 
-  this.parent = parent;
-  if (this.parent) {
-    this.prefix = parent.prefix + this.prefix;
+    const triggersExit = options.triggersExit || [];
+    options.triggersExit = triggersExit.concat(this._triggersExit);
 
-    this._triggersEnter = parent._triggersEnter.concat(this._triggersEnter);
-    this._triggersExit = this._triggersExit.concat(parent._triggersExit);
-  }
-};
-
-Group.prototype.route = function(pathDef, options, group) {
-  options = options || {};
-
-  if (!/^\/.*/.test(pathDef)) {
-    var message = "route's path must start with '/'";
-    throw new Error(message);
+    return super.route(pathDef, options, group);
   }
 
-  group = group || this;
-  pathDef = this.prefix + pathDef;
+  callSubscriptions(current) {
+    if (this.parent) {
+      this.parent.callSubscriptions(current);
+    }
 
-  var triggersEnter = options.triggersEnter || [];
-  options.triggersEnter = this._triggersEnter.concat(triggersEnter);
-
-  var triggersExit = options.triggersExit || [];
-  options.triggersExit = triggersExit.concat(this._triggersExit);
-
-  return this._router.route(pathDef, options, group);
-};
-
-Group.prototype.group = function(options) {
-  return new Group(this._router, options, this);
-};
-
-Group.prototype.callSubscriptions = function(current) {
-  if (this.parent) {
-    this.parent.callSubscriptions(current);
+    this._subscriptions.call(current.route, current.params, current.queryParams);
   }
-
-  this._subscriptions.call(current.route, current.params, current.queryParams);
 };