+++ package.js
@@ -1,15 +1,18 @@
 Package.describe({
-  name: 'kadira:flow-router',
-  summary: 'Carefully Designed Client Side Router for Meteor',
-  version: '2.10.0',
+  name: 'kadira:flow-router-ssr',
+  summary: 'Same as kadira:flow-router, but with SSR support',
+  version: '3.7.3',
   git: 'https://github.com/kadirahq/flow-router.git'
 });
 
 Npm.depends({
+  deepmerge: '0.2.10',
+  'cookie-parser': '1.4.0',
+  cheerio: '0.19.0',
   // In order to support IE9, we had to fork pagejs and apply
   // this PR: https://github.com/visionmedia/page.js/pull/288
-  'page':'https://github.com/kadirahq/page.js/archive/34ddf45ea8e4c37269ce3df456b44fc0efc595c6.tar.gz',
-  'qs':'5.2.0'
+  page: 'https://github.com/kadirahq/page.js/archive/34ddf45ea8e4c37269ce3df456b44fc0efc595c6.tar.gz',
+  qs: '5.2.0'
 });
 
 Package.onUse(function(api) {
@@ -22,14 +25,14 @@ Package.onTest(function(api) {
   api.use('tinytest');
   api.use('check');
   api.use('mongo');
+  api.use('minimongo');
   api.use('http');
   api.use('random');
   api.use('practicalmeteor:sinon');
   api.use('meteorhacks:fast-render');
   api.use('meteorhacks:inject-data');
   api.use('tmeasday:html5-history-api');
-
-  api.addFiles('test/common/fast_render_route.js', ['client', 'server']);
+  api.use('smithy:describe@1.0.0');
 
   api.addFiles('test/client/_helpers.js', 'client');
   api.addFiles('test/server/_helpers.js', 'server');
@@ -37,46 +40,57 @@ Package.onTest(function(api) {
   api.addFiles('test/client/loader.spec.js', 'client');
   api.addFiles('test/client/route.reactivity.spec.js', 'client');
   api.addFiles('test/client/router.core.spec.js', 'client');
-  api.addFiles('test/client/router.subs_ready.spec.js', 'client');
   api.addFiles('test/client/router.reactivity.spec.js', 'client');
   api.addFiles('test/client/group.spec.js', 'client');
   api.addFiles('test/client/trigger.spec.js', 'client');
   api.addFiles('test/client/triggers.js', 'client');
 
-  api.addFiles('test/server/plugins/fast_render.js', 'server');
-
+  api.addFiles('test/common/loader.spec.js', ['client', 'server']);
   api.addFiles('test/common/router.path.spec.js', ['client', 'server']);
   api.addFiles('test/common/router.url.spec.js', ['client', 'server']);
   api.addFiles('test/common/router.addons.spec.js', ['client', 'server']);
   api.addFiles('test/common/route.spec.js', ['client', 'server']);
   api.addFiles('test/common/group.spec.js', ['client', 'server']);
+
+  api.addFiles('server/__tests__/ssr_context.js', 'server');
+  api.addFiles('server/__tests__/route.js', 'server');
+  api.addFiles('server/__tests__/group.js', 'server');
+  api.addFiles('server/plugins/__tests__/ssr_data.js', 'server');
 });
 
 function configure(api) {
-  api.versionsFrom('1.0');
+  api.versionsFrom('1.2');
 
+  api.use('ecmascript');
   api.use('underscore');
   api.use('tracker');
   api.use('reactive-dict');
   api.use('reactive-var');
+  api.use('ddp');
   api.use('ejson');
-
-  api.use('meteorhacks:fast-render@2.10.0', ['client', 'server'], {weak: true});
+  api.use('meteorhacks:fast-render@2.11.0', ['client', 'server']);
   api.use('cosmos:browserify@0.9.2', 'client');
+  api.use('meteorhacks:picker@1.0.3', 'server');
+  api.use('meteorhacks:inject-data@1.4.1');
+
+  api.addFiles('lib/router.js', ['client', 'server']);
+  api.addFiles('lib/group.js', ['client', 'server']);
+  api.addFiles('lib/route.js', ['client', 'server']);
 
   api.addFiles('client.browserify.js', 'client');
   api.addFiles('client/triggers.js', 'client');
   api.addFiles('client/router.js', 'client');
   api.addFiles('client/group.js', 'client');
   api.addFiles('client/route.js', 'client');
-  api.addFiles('client/_init.js', 'client');
 
   api.addFiles('server/router.js', 'server');
   api.addFiles('server/group.js', 'server');
   api.addFiles('server/route.js', 'server');
+  api.addFiles('server/ssr_context.js', 'server');
+
+  api.addFiles('lib/_init.js', ['client', 'server']);
+  api.addFiles('client/_init.js', 'client');
   api.addFiles('server/_init.js', 'server');
 
-  api.addFiles('server/plugins/fast_render.js', 'server');
-
-  api.addFiles('lib/router.js', ['client', 'server']);
+  api.addFiles('server/plugins/ssr_data.js', 'server');
 }