+++ test/common/router.addons.spec.js
@@ -1,20 +1,26 @@
 Router = FlowRouter.Router;
 
-Tinytest.addAsync('Common - Addons - onRouteRegister basic usage', function (test, done) {
+Tinytest.addAsync('Common - Addons - onRouteRegister basic usage', function(test, done) {
   var name = Random.id();
   var customField = Random.id();
   var pathDef = '/' + name;
-  
+
   FlowRouter.onRouteRegister(function(route) {
     test.equal(route, {
       pathDef: pathDef,
-
       // Route.path is deprecated and will be removed in 3.0
       path: pathDef,
-
       name: name,
       options: {customField: customField}
-    });  
+    });
+
+    test.equal(route.pathDef, pathDef);
+    // Route.path is deprecated and will be removed in 3.0
+    test.equal(route.path, pathDef);
+    test.equal(route.name, name);
+    test.equal(route.options.customField, customField);
+    test.equal(route.options, {customField: customField});
+
     FlowRouter._onRouteCallbacks = [];
     done();
   });
@@ -22,7 +28,6 @@ Tinytest.addAsync('Common - Addons - onRouteRegister basic usage', function (tes
   FlowRouter.route(pathDef, {
     name: name,
     action: function() {},
-    subscriptions: function() {},
     triggersEnter: function() {},
     triggersExit: function() {},
     customField: customField